+++ dejafu/Test/DejaFu/Utils.hs
@@ -85,10 +85,7 @@ showFail :: Failure -> String
 showFail Abort = "[abort]"
 showFail Deadlock = "[deadlock]"
 showFail STMDeadlock = "[stm-deadlock]"
-showFail InternalError = "[internal-error]"
 showFail (UncaughtException exc) = "[" ++ displayException exc ++ "]"
-showFail IllegalSubconcurrency = "[illegal-subconcurrency]"
-showFail IllegalDontCheck = "[illegal-dontcheck]"
 
 -------------------------------------------------------------------------------
 -- * Scheduling